+++ pkg/edgeworkers/edgekv_namespaces.go
@@ -6,8 +6,9 @@ import (
 	"fmt"
 	"net/http"
 	"net/url"
+	"time"
 
-	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v10/pkg/session"
+	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v11/pkg/session"
 	validation "github.com/go-ozzo/ozzo-validation/v4"
 )
 
@@ -18,7 +19,7 @@ type (
 		Details bool
 	}
 
-	// GetEdgeKVNamespaceRequest contains path parameters used to feth a namespace
+	// GetEdgeKVNamespaceRequest contains path parameters used to fetch a namespace
 	GetEdgeKVNamespaceRequest struct {
 		Network NamespaceNetwork
 		Name    string
@@ -58,6 +59,25 @@ type (
 
 	// NamespaceNetwork represents available namespace network types
 	NamespaceNetwork string
+
+	// DeleteEdgeKVNamespaceRequest represents the request to delete a namespace.
+	DeleteEdgeKVNamespaceRequest struct {
+		// Network specifies the network environment to execute the API request on.
+		Network NamespaceNetwork
+
+		// Name is a unique identifier for each namespace.
+		Name string
+
+		// Sync specifies whether to delete the namespace synchronously or asynchronously.
+		Sync bool
+	}
+
+	// DeleteEdgeKVNamespacesResponse represents a response object returned when deleting a namespace.
+	DeleteEdgeKVNamespacesResponse struct {
+		// ScheduledDeleteTime is the time when the namespace will be deleted for asynchronous deletion.
+		// For synchronous deletion, it will be nil.
+		ScheduledDeleteTime *time.Time `json:"scheduledDeleteTime"`
+	}
 )
 
 const (
@@ -106,6 +126,15 @@ func (r UpdateEdgeKVNamespaceRequest) Validate() error {
 	}.Filter()
 }
 
+// Validate validates DeleteEdgeKVNamespaceRequest
+func (r DeleteEdgeKVNamespaceRequest) Validate() error {
+	return validation.Errors{
+		"Network": validation.Validate(r.Network, validation.Required, validation.In(NamespaceStagingNetwork, NamespaceProductionNetwork).Error(
+			fmt.Sprintf("value '%s' is invalid. Must be one of: '%s' or '%s'", r.Network, NamespaceStagingNetwork, NamespaceProductionNetwork))),
+		"Name": validation.Validate(r.Name, validation.Required, validation.Length(1, 32)),
+	}.Filter()
+}
+
 func validateRetention(value interface{}) error {
 	v, ok := value.(*int)
 	if !ok {
@@ -143,6 +172,8 @@ var (
 	ErrCreateEdgeKVNamespace = errors.New("create an EdgeKV namespace")
 	// ErrUpdateEdgeKVNamespace is returned when UpdateEdgeKVNamespace fails
 	ErrUpdateEdgeKVNamespace = errors.New("update an EdgeKV namespace")
+	// ErrDeleteEdgeKVNamespace is returned when DeleteEdgeKVNamespace fails
+	ErrDeleteEdgeKVNamespace = errors.New("delete an EdgeKV namespace")
 )
 
 func (e *edgeworkers) ListEdgeKVNamespaces(ctx context.Context, params ListEdgeKVNamespacesRequest) (*ListEdgeKVNamespacesResponse, error) {
@@ -262,6 +293,51 @@ func (e *edgeworkers) UpdateEdgeKVNamespace(ctx context.Context, params UpdateEd
 
 	if resp.StatusCode != http.StatusOK {
 		return nil, fmt.Errorf("%s: %w", ErrUpdateEdgeKVNamespace, e.Error(resp))
+	}
+
+	return &result, nil
+}
+
+func (e *edgeworkers) DeleteEdgeKVNamespace(ctx context.Context, params DeleteEdgeKVNamespaceRequest) (*DeleteEdgeKVNamespacesResponse, error) {
+	logger := e.Log(ctx)
+	logger.Debug("DeleteEdgeKVNamespace")
+
+	if err := params.Validate(); err != nil {
+		return nil, fmt.Errorf("%s: %w: %s", ErrDeleteEdgeKVNamespace, ErrStructValidation, err.Error())
+	}
+
+	uri, err := url.Parse(fmt.Sprintf("/edgekv/v1/networks/%s/namespaces/%s", params.Network, params.Name))
+	if err != nil {
+		return nil, fmt.Errorf("%w: failed to parse url: %s", ErrDeleteEdgeKVNamespace, err.Error())
+	}
+
+	if params.Sync {
+		q := url.Values{}
+		q.Add("sync", "true")
+		uri.RawQuery = q.Encode()
+	}
+
+	req, err := http.NewRequestWithContext(ctx, http.MethodDelete, uri.String(), nil)
+	if err != nil {
+		return nil, fmt.Errorf("%w: failed to create request: %s", ErrDeleteEdgeKVNamespace, err.Error())
+	}
+
+	var result DeleteEdgeKVNamespacesResponse
+	resp, err := e.Exec(req, &result)
+	if err != nil {
+		return nil, fmt.Errorf("%w: request failed: %s", ErrDeleteEdgeKVNamespace, err.Error())
+	}
+	defer session.CloseResponseBody(resp)
+
+	var expectedCode int
+	if params.Sync {
+		expectedCode = http.StatusOK
+	} else {
+		expectedCode = http.StatusAccepted
+	}
+
+	if resp.StatusCode != expectedCode {
+		return nil, fmt.Errorf("%s: %w", ErrDeleteEdgeKVNamespace, e.Error(resp))
 	}
 
 	return &result, nil