+++ pkg/iam/api_clients.go
@@ -9,8 +9,8 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v10/pkg/edgegriderr"
-	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v10/pkg/session"
+	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v11/pkg/edgegriderr"
+	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v11/pkg/session"
 	validation "github.com/go-ozzo/ozzo-validation/v4"
 )
 
@@ -114,11 +114,10 @@ type (
 		CreatedDate             time.Time                   `json:"createdDate"`
 		Credentials             []CreateAPIClientCredential `json:"credentials"`
 		GroupAccess             GroupAccess                 `json:"groupAccess"`
-		IPACL                   IPACL                       `json:"ipAcl"`
+		IPACL                   *IPACL                      `json:"ipAcl"`
 		IsLocked                bool                        `json:"isLocked"`
 		NotificationEmails      []string                    `json:"notificationEmails"`
-		PurgeOptions            PurgeOptions                `json:"purgeOptions"`
-		ServiceProviderID       int64                       `json:"serviceProviderId"`
+		PurgeOptions            *PurgeOptions               `json:"purgeOptions"`
 	}
 
 	// GetAPIClientResponse describes the response of the GetAPIClient endpoint.
@@ -139,11 +138,10 @@ type (
 		CreatedDate             time.Time             `json:"createdDate"`
 		Credentials             []APIClientCredential `json:"credentials"`
 		GroupAccess             GroupAccess           `json:"groupAccess"`
-		IPACL                   IPACL                 `json:"ipAcl"`
+		IPACL                   *IPACL                `json:"ipAcl"`
 		IsLocked                bool                  `json:"isLocked"`
 		NotificationEmails      []string              `json:"notificationEmails"`
-		PurgeOptions            PurgeOptions          `json:"purgeOptions"`
-		ServiceProviderID       int64                 `json:"serviceProviderId"`
+		PurgeOptions            *PurgeOptions         `json:"purgeOptions"`
 	}
 
 	// APIClientActions specifies activities available for the API client.
@@ -154,7 +152,7 @@ type (
 		EditAPIs          bool `json:"editApis"`
 		EditAuth          bool `json:"editAuth"`
 		EditGroups        bool `json:"editGroups"`
-		EditIPAcl         bool `json:"editIpAcl"`
+		EditIPACL         bool `json:"editIpAcl"`
 		EditSwitchAccount bool `json:"editSwitchAccount"`
 		Lock              bool `json:"lock"`
 		Transfer          bool `json:"transfer"`
@@ -169,12 +167,24 @@ type (
 
 	// API represents single Application Programming Interface (API).
 	API struct {
-		AccessLevel      AccessLevel `json:"accessLevel"`
-		APIID            int64       `json:"apiId"`
-		APIName          string      `json:"apiName"`
-		Description      string      `json:"description"`
-		DocumentationURL string      `json:"documentationUrl"`
-		Endpoint         string      `json:"endPoint"`
+		AccessLevel      string `json:"accessLevel"`
+		APIID            int64  `json:"apiId"`
+		APIName          string `json:"apiName"`
+		Description      string `json:"description"`
+		DocumentationURL string `json:"documentationUrl"`
+		Endpoint         string `json:"endPoint"`
+	}
+
+	// APIAccessRequest represents the APIs the API client can access.
+	APIAccessRequest struct {
+		AllAccessibleAPIs bool             `json:"allAccessibleApis"`
+		APIs              []APIRequestItem `json:"apis"`
+	}
+
+	// APIRequestItem represents single Application Programming Interface (API).
+	APIRequestItem struct {
+		APIID       int64       `json:"apiId"`
+		AccessLevel AccessLevel `json:"accessLevel"`
 	}
 
 	// APIClientCredential represents single Credential returned by APIClient interfaces.
@@ -218,6 +228,19 @@ type (
 		Subgroups       []ClientGroup `json:"subgroups"`
 	}
 
+	// GroupAccessRequest specifies the API client's group access.
+	GroupAccessRequest struct {
+		CloneAuthorizedUserGroups bool                     `json:"cloneAuthorizedUserGroups"`
+		Groups                    []ClientGroupRequestItem `json:"groups"`
+	}
+
+	// ClientGroupRequestItem represents a group the API client can access.
+	ClientGroupRequestItem struct {
+		GroupID   int64                    `json:"groupId"`
+		RoleID    int64                    `json:"roleId"`
+		Subgroups []ClientGroupRequestItem `json:"subgroups"`
+	}
+
 	// IPACL specifies the API client's IP list restriction.
 	IPACL struct {
 		CIDR   []string `json:"cidr"`
@@ -239,18 +262,18 @@ type (
 
 	// CreateAPIClientRequest contains the request parameters for the CreateAPIClient endpoint.
 	CreateAPIClientRequest struct {
-		AllowAccountSwitch      bool          `json:"allowAccountSwitch"`
-		APIAccess               APIAccess     `json:"apiAccess"`
-		AuthorizedUsers         []string      `json:"authorizedUsers"`
-		CanAutoCreateCredential bool          `json:"canAutoCreateCredential"`
-		ClientDescription       string        `json:"clientDescription"`
-		ClientName              string        `json:"clientName"`
-		ClientType              ClientType    `json:"clientType"`
-		CreateCredential        bool          `json:"createCredential"`
-		GroupAccess             GroupAccess   `json:"groupAccess"`
-		IPACL                   *IPACL        `json:"ipAcl,omitempty"`
-		NotificationEmails      []string      `json:"notificationEmails"`
-		PurgeOptions            *PurgeOptions `json:"purgeOptions,omitempty"`
+		AllowAccountSwitch      bool               `json:"allowAccountSwitch"`
+		APIAccess               APIAccessRequest   `json:"apiAccess"`
+		AuthorizedUsers         []string           `json:"authorizedUsers"`
+		CanAutoCreateCredential bool               `json:"canAutoCreateCredential"`
+		ClientDescription       string             `json:"clientDescription"`
+		ClientName              string             `json:"clientName"`
+		ClientType              ClientType         `json:"clientType"`
+		CreateCredential        bool               `json:"createCredential"`
+		GroupAccess             GroupAccessRequest `json:"groupAccess"`
+		IPACL                   *IPACL             `json:"ipAcl,omitempty"`
+		NotificationEmails      []string           `json:"notificationEmails"`
+		PurgeOptions            *PurgeOptions      `json:"purgeOptions,omitempty"`
 	}
 
 	// UpdateAPIClientRequest contains the request parameters for the UpdateAPIClient endpoint.
@@ -261,17 +284,17 @@ type (
 
 	// UpdateAPIClientRequestBody represents body params for the UpdateAPIClient endpoint.
 	UpdateAPIClientRequestBody struct {
-		AllowAccountSwitch      bool          `json:"allowAccountSwitch"`
-		APIAccess               APIAccess     `json:"apiAccess"`
-		AuthorizedUsers         []string      `json:"authorizedUsers"`
-		CanAutoCreateCredential bool          `json:"canAutoCreateCredential"`
-		ClientDescription       string        `json:"clientDescription"`
-		ClientName              string        `json:"clientName"`
-		ClientType              ClientType    `json:"clientType"`
-		GroupAccess             GroupAccess   `json:"groupAccess"`
-		IPACL                   *IPACL        `json:"ipAcl,omitempty"`
-		NotificationEmails      []string      `json:"notificationEmails"`
-		PurgeOptions            *PurgeOptions `json:"purgeOptions,omitempty"`
+		AllowAccountSwitch      bool               `json:"allowAccountSwitch"`
+		APIAccess               APIAccessRequest   `json:"apiAccess"`
+		AuthorizedUsers         []string           `json:"authorizedUsers"`
+		CanAutoCreateCredential bool               `json:"canAutoCreateCredential"`
+		ClientDescription       string             `json:"clientDescription"`
+		ClientName              string             `json:"clientName"`
+		ClientType              ClientType         `json:"clientType"`
+		GroupAccess             GroupAccessRequest `json:"groupAccess"`
+		IPACL                   *IPACL             `json:"ipAcl,omitempty"`
+		NotificationEmails      []string           `json:"notificationEmails"`
+		PurgeOptions            *PurgeOptions      `json:"purgeOptions,omitempty"`
 	}
 
 	// UpdateAPIClientResponse describes the response from the UpdateAPIClient endpoint.
@@ -291,6 +314,12 @@ const (
 	ReadWriteLevel AccessLevel = "READ-WRITE"
 	// ReadOnlyLevel is the `READ-ONLY` access level.
 	ReadOnlyLevel AccessLevel = "READ-ONLY"
+	// ReadLevel is the `READ` access level.
+	ReadLevel AccessLevel = "READ"
+	// CredentialReadOnlyLevel is the `CREDENTIAL-READ-ONLY` access level.
+	CredentialReadOnlyLevel AccessLevel = "CREDENTIAL-READ-ONLY"
+	// CredentialReadWriteLevel is the `CREDENTIAL-READ-WRITE` access level.
+	CredentialReadWriteLevel AccessLevel = "CREDENTIAL-READ-WRITE"
 )
 
 // Validate validates UnlockAPIClientRequest.
@@ -311,31 +340,32 @@ func (r CreateAPIClientRequest) Validate() error {
 	})
 }
 
-// Validate validates APIAccess.
-func (a APIAccess) Validate() error {
+// Validate validates APIAccessRequest.
+func (a APIAccessRequest) Validate() error {
 	return validation.Errors{
 		"APIs": validation.Validate(a.APIs, validation.When(!a.AllAccessibleAPIs, validation.Required)),
 	}.Filter()
 }
 
-// Validate validates API.
-func (a API) Validate() error {
+// Validate validates APIRequest.
+func (a APIRequestItem) Validate() error {
 	return validation.Errors{
-		"AccessLevel": validation.Validate(a.AccessLevel, validation.Required, validation.In(ReadOnlyLevel, ReadWriteLevel).Error(
-			fmt.Sprintf("value '%s' is invalid. Must be one of: 'READ-ONLY' or 'READ-WRITE'", a.AccessLevel))),
+		"AccessLevel": validation.Validate(a.AccessLevel, validation.Required,
+			validation.In(ReadOnlyLevel, ReadWriteLevel, ReadLevel, CredentialReadOnlyLevel, CredentialReadWriteLevel).Error(
+				fmt.Sprintf("value '%s' is invalid. Must be one of: 'READ-ONLY', 'READ-WRITE', 'READ', 'CREDENTIAL-READ-ONLY' or 'CREDENTIAL-READ-WRITE'", a.AccessLevel))),
 		"APIID": validation.Validate(a.APIID, validation.Required),
 	}.Filter()
 }
 
-// Validate validates GroupAccess.
-func (ga GroupAccess) Validate() error {
+// Validate validates GroupAccessRequest.
+func (ga GroupAccessRequest) Validate() error {
 	return validation.Errors{
 		"Groups": validation.Validate(ga.Groups, validation.When(!ga.CloneAuthorizedUserGroups, validation.Required)),
 	}.Filter()
 }
 
-// Validate validates ClientGroup.
-func (cg ClientGroup) Validate() error {
+// Validate validates ClientGroupRequest.
+func (cg ClientGroupRequestItem) Validate() error {
 	return validation.Errors{
 		"GroupID": validation.Validate(cg.GroupID, validation.Required),
 		"RoleID":  validation.Validate(cg.RoleID, validation.Required),