+++ pkg/iam/helper.go
@@ -8,8 +8,8 @@ import (
 	"net/url"
 	"strconv"
 
-	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v10/pkg/edgegriderr"
-	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v10/pkg/session"
+	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v11/pkg/edgegriderr"
+	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v11/pkg/session"
 	validation "github.com/go-ozzo/ozzo-validation/v4"
 )
 
@@ -34,8 +34,8 @@ type (
 
 	// ListAllowedCPCodesRequestBody contains the filtering parameters for the ListAllowedCPCodes endpoint.
 	ListAllowedCPCodesRequestBody struct {
-		ClientType ClientType            `json:"clientType"`
-		Groups     []AllowedCPCodesGroup `json:"groups"`
+		ClientType ClientType               `json:"clientType"`
+		Groups     []ClientGroupRequestItem `json:"groups"`
 	}
 
 	// AllowedCPCodesGroup contains the group parameters for the ListAllowedCPCodes endpoint.